Make `BlockHeightUnsafe` wrap `ULong` instead of `Long`
Block heights are represented as `uint64` in all lightwalletd response
types, so they should be parsed as ULong instead of Long values.

+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/model/ext/BlockHeightUnsafeExt.kt
@@ -3,6 +3,6 @@ package cash.z.ecc.android.sdk.internal.model.ext
 import cash.z.ecc.android.sdk.model.BlockHeight
 import co.electriccoin.lightwallet.client.model.BlockHeightUnsafe
 
-internal fun BlockHeightUnsafe.Companion.from(blockHeight: BlockHeight) = BlockHeightUnsafe(blockHeight.value)
+internal fun BlockHeightUnsafe.Companion.from(blockHeight: BlockHeight) = BlockHeightUnsafe(blockHeight.value.toULong())
 
-internal fun BlockHeightUnsafe.toBlockHeight() = BlockHeight.new(value)
+internal fun BlockHeightUnsafe.toBlockHeight() = BlockHeight.new(value.toLong())
